backend/requirements.txt
openai>=1.50.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
pillow>=11.0.0
httpx>=0.27.0
pydantic>=2.9.0
